+++ shelf_packages/open_api_specification/lib/src/organize_open_api.dart
@@ -32,8 +32,9 @@ Map<String, dynamic> organizeOpenApi(Map<String, dynamic> document) {
             return MapEntry(
               type,
               media?.map((key, value) {
-                if (key == 'schema')
+                if (key == 'schema') {
                   return MapEntry(key, resolveSchema(value as Map<String, dynamic>));
+                }
                 return MapEntry(key, value);
               }),
             );