+++ shelf_packages/shelf_open_api_generator/example/lib/features/messages/dto/message_fetch_dto.g.dart
@@ -1,5 +1,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: unnecessary_raw_strings
+
 part of 'message_fetch_dto.dart';
 
 // **************************************************************************