+++ shelf_packages/shelf_open_api_generator/lib/src/utils/json_annotation.dart
@@ -1,19 +1,19 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:source_gen/source_gen.dart';
 
 abstract final class JsonAnnotation {
-  static const _keyChecker = TypeChecker.fromRuntime(JsonKey);
-  static const _valueChecker = TypeChecker.fromRuntime(JsonValue);
+  static const _keyChecker = TypeChecker.typeNamed(JsonKey, inPackage: 'json_annotation');
+  static const _valueChecker = TypeChecker.typeNamed(JsonValue, inPackage: 'json_annotation');
 
-  static String? getFieldName(Element element) {
+  static String? getFieldName(Element2 element) {
     final annotation = _keyChecker.firstAnnotationOf(element);
     if (annotation == null) return null;
 
     return annotation.getField('name')?.toStringValue();
   }
 
-  static Object? getEnumValue(Element element) {
+  static Object? getEnumValue(Element2 element) {
     final annotation = _valueChecker.firstAnnotationOf(element);
     if (annotation == null) return null;