+++ shelf_packages/shelf_routing_generator/lib/src/route_header_handler.dart
@@ -1,15 +1,15 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:shelf_routing/shelf_routing.dart';
 import 'package:source_gen/source_gen.dart';
 
 class RouteHeaderHandler {
-  static const TypeChecker _checker = TypeChecker.fromRuntime(RouteHeader);
+  static const _checker = TypeChecker.typeNamed(RouteHeader, inPackage: 'shelf_routing');
 
   final String name;
 
   const RouteHeaderHandler._({required this.name});
 
-  static List<RouteHeaderHandler> from(MethodElement element) {
+  static List<RouteHeaderHandler> from(MethodElement2 element) {
     return _checker.annotationsOf(element).map(ConstantReader.new).map((e) {
       return RouteHeaderHandler._(name: e.read('name').stringValue);
     }).toList();