+++ shelf_packages/open_api_specification/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.0.0
+- build: bumped dart sdk version to `^3.8.0`
 - fix!: fixed security field in open api operation
 
 ## 1.0.0